Fix issue that broke all the shell benchmarks
Apparently the benchmarker was quite brittle and broke unexpectedly in
one of my updates yesterday.
I spend hours trying to find what the hell happend, and it seems that
the parameters were not interpretd well. Arff such a pain.
Now I fixed most of them but something else broke. So I decided to make
this commit and then jump in for the next problem