Fix python benchmarking!
It makes sense that the hashlib implementation is about as fast as the
OpenSSL implementation, as it seems the library uses openSSL for that.

So for a few weeks I was very surpised that the python impl would
perform better than all the c++ counterparts, and I had pending to come
back to it and look closer. Finally I did, and for my surprise I saw
that the shell script was playing me a trick: when I executed pyperf on
the commandline the timings matched with my expectations. However when I
ran this on the script the timings were too good to be true.
I had to do a few tries to find out how to fix the shell script so that
it did the same as in the command line. Now I can say it works.

Just a pity I had the wrong data for a few weeks. But because of that I
got to investigate and find other alternatives I'd like to benchmark
here, among them using Intel's SHA-NI extensions (that very likely be
the fastest out there).